requirements.txt
flask
gunicorn
openai
python-dotenv
notion-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client